fix: Always update build.gradle and pom.xml
When installing the agent in a Java project, always update the build
configuration with the latest spec for the agent. This matches the
behavior in a Ruby project, and makes it easier to update a Java
project.

Also, these changes ensure that the "plugins" block will be situated
correctly in build.gradle (i.e. directly after the "buildscript" block),
and add tests for installing in a Java project.

Finally, bump the versions of the gradle and maven plugins to 1.1.0 and
1.2.0, respectively.