refactor: Separate user input from installer
Installers now implement the AgentInstaller interface. A language can
have multiple installers for various frameworks or project types, and
are summarized as the AgentInstallerOptions interface. A new class,
AgentInstallerProcedure, takes one of these options objects and uses it
to ask the user questions and perform the install through a generic
interface.