feat: Rework the `agent-install` user experience (#316)
* Migrate some file extensions from .js -> .ts

* Typescriptify

* refactor: Separate user input from installer

Installers now implement the AgentInstaller interface. A language can
have multiple installers for various frameworks or project types, and
are summarized as the AgentInstallerOptions interface. A new class,
AgentInstallerProcedure, takes one of these options objects and uses it
to ask the user questions and perform the install through a generic
interface.

* Command output and details are logged when running with verbose

* fix: plugin section mustcome before other declarations

* fix: Prompt to add `mavenCentral` repository when installing the gradle
plugin

* fix: Append a version if it was previously missing

* fix: Respect whitespace when editing build.gradle

* clean up error messages

* allow user to supply project framework in addition to language

Instead of installing for `java`, the user can now specify the framework
which also implies the language, e.g. `maven` implies `java`.

* style: move install-agent require to the top

* test: refactor tests to work with new interfaces

* update dependencies

* warn user of unsupported configuration without java plugin

* Migrate cli.js to TypeScript

This fixes compiler errors and makes no attempt at moving toward
idiomatic TypeScript.

* Don't rely on OS when determining the wrapper to use

This may fail in some cases on Windows such as when running in Cygwin

* Don't end copy+pasteable text with a period (user feedback)

* remove hardcoded documentation link