feat: Improve user feedback for `view source`
- We now display the source location for classes, functions and events.
- If the source can be opened, an external link icon is now present
- If the source cannot be opened for any reason, the owner of the view
has the ability to add a warning message below the source code
describing the issue.

BREAKING CHANGE: This change alters the `viewSource` event emit by the
VsCodeExtension view. It now requires the owner listen for and respond
to a `request-resolve-location` event, providing a raw location to be
transformed and sent back in a `response-resolve-location` event.