fix: Indexing no longer causes EXDEV on Linux
The code tried really hard to do atomic directory switch, which
in addition to being impossible caused an error when tmpdir was
on a different volume (as is common on Linux).

Instead, new code just writes files atomically and makes sure to
write mtime last. Downstream code should watch for changes
to that file to detect when reindexing occurs.