fix: Reliably determine that an AppMap index is up-to-date
First, don't use ctime to determine when a file is created -
because that's not what it does. In fact, it's:

Time when file status was last changed (inode data modification).

Second, move the normalized AppMap file into place after writing the
mtime file, so that when the indexer is triggered by the new/updated
AppMap file, it will always find an mtime file that tells it that the
index is already up-to-date.