feat: Track specific untrusted data in unsafe deserialization rule
Unsafe deserialization scanner rule now tracks untrusted data
through the call tree instead of relying instead on trusted sources
being tagged and only checking the call sequence.

This means that common false positives are less likely (such as
for example arising from unmarshaling trusted on-disk data, as is
frequent in some frameworks).

Since specific pieces of data are now tracked some false negatives
can also be avoided, for example when multiple tainted data inputs
are used but only some of them are sanitized.

The disadvantage is that the reliability of the rule now depends
on completeness of the AppMap. If there is a data transformation
that is not captured it's invisible to the rule and will result
in failure to associate it with the tracked untrusted data.

Note also that currently the tracked tainted data is limited to
HTTP request messages, although there is no reason this cannot
be expanded in the future.