perf: Don't try to index the same AppMaps more than once
When files are changing quickly, new and changed AppMaps could
be added more than once to the processing queue. This is relatively
harmless since the process would check the index for freshness
and bail early, but still this added some unneccessary extra work.
This has been exacerbated by the recent addition of a poller to
fix watching reliability.

To avoid this issue, keep a list of pending AppMaps and don't add
it to the queue if it's already on the list.