feat: remove fonky default test command
The default js test command seems to record 0 appmaps in many cases. It 
could be that there is no tests or that the project is not using node to 
run tests. Also, the command is confusing because it mixes 
`--recorder=process` with a test command whereas we want to push ad-hoc 
recorders for tests.