fix: fix event highlighting in trace view
This fixes a bug that occurs when trying to view a highlighted event in
the trace view. If the event was accessed through the `Related Events`
section of the finding details panel, then switching to the trace view
would not properly highlight the event because the event was not derived
from the `filteredAppMap` computed property that was used to generate the
trace view.