fix: Allow non-standard requirements.txt names
When there's only a single requirements.txt file in a Python
project make sure to use its name instead of assuming it's always
'requirements.txt'.

Also, change the tests for this code to only excercise this code path
instead of trying to call out to python, which was slow and unreliable.