Skip type checking in scanner schema generator
I'm not sure why it started happening with the addition of
cli->client dependency (doesn't make any sense, it's not related
to scanner at all!), but the following error cropped up:

    node_modules/@types/eslint-scope/node_modules/@types/eslint/helpers.d.ts(1,6): error TS2300: Duplicate identifier 'Prepend'.
    node_modules/@types/eslint/helpers.d.ts(1,6): error TS2300: Duplicate identifier 'Prepend'.

Since we're type checking this project in other places anyway,
there's no need for the schema generator to do that too. Disabling
the check fixes this issue.