fix: Remove side effect in `classNameToOpenAPIType`
When given an unknown class, the first time the method was called it
would return 'object'. Subsequent calls for that class would return
`undefined`.

Instead, it'll always return one or the other depending on the options
provided. Setting `strict` to true will return `undefined` if the class
is not known while leaving it falsy will cause the function to return
`object`. This is due to the fact that in some cases there was code
relying on it returning `undefined`, while other cases it relied on it
returning `object`.