feat: update prune cli command
Previously, the prune command would remove the most expensive events
from each 'chunk' of the appmap, but this made it difficult to notify
the user what events were pruned. This change makes it so that the most
expensive events for the whole appmap are removed. Also, the --format
option is added to the prune command.