test: Use standard Jest method of suppressing log messages
There's no need to use Sinon to mock console; Jest already controls
the console and provides a way to silence the messages in tests.

This change was needed also because using Sinon to stub console like
that was problematic with AppMap. Using Jest's built-in spies works
perfectly fine. It's also cleaner IMO.