build: fix dynamic import to get rid of hacky rollup replace
Please, review this carefully.

I think the problem was that dynamic import returns the entire
namespace of modules. And we only want the default export. So I changed
the dynamic module resolution (which was introduced to solve a circular
dependency) to return the default member. IMO this is much better than
the hacky rollup replace.