build: add inlineDynamicImports in rollup config
Please, review this carefully.

This is to fix the following rollup error when running `yarn build` on
the upcoming version of the flamegraph:

```
[!] Error: Invalid value for option "output.file" - when building multiple chunks, the "output.dir" option must be used, not "output.file". To inline dynamic imports, set the "inlineDynamicImports" option.
```

The upcoming version of the flamegraph has circular dependency between
components so it needs to break the cycle be using dynamic import.
However, I don't understand why this error did not appear before because
we already had dynamic import in `trace/TraceEventBlock.vue` [1].

[1]: https://github.com/getappmap/appmap-js/blob/8603c7134bb76f0a4c62f7b3000dcf174dc61972/packages/components/src/components/trace/TraceEventBlock.vue#L57