refactor: re-implement zoom slider as vue component
The original implementation of the zoom slider is:
https://github.com/getappmap/appmap-js/blob/main/packages/diagrams/src/helpers/container/zoom.js

The problem with reusing the original implementation is that it is
entangled with functionalities like drag and drop panel, which we don't
need here. Also it is not a vue component, so it is less consistent with
the packages/components codebase. In the future this component could be
reused to implement the zoom slider of the dependency diagram and trace
diagram.