feat: zoom centering and partial scrolling
Sorry this commit has two changes which was hard to disentangle. It
resulted into the creation of a new component `FlamegraphMain` which
handles scrolling and zooming.

First partial scrolling. It did not make sense to scroll the hover label
and the root. Now, only actual flamegraph items are scrollable.

Second, zoom centering. This caused me a lot of pain but the solution
was simple in the end. We basically want to keep a point in
flamegraph (the origin) immobile during zooming. The problem was to
combine programmatic scrolling with css width animation. While the
initial and final state are easy to compute, I could not make the
transition smoooth. This became impossibly jumpy on higher zoom level.
The solution was simple: only enable css animation while focusing and
let the high changing rate of the zoom dictate the animation.

There two mode of zooming:
- When the mouse is outside the window, we zoom on the center. This
  happens when zomming with the slider.
- When the mouse is inside the window, we zoom on the mouse position.
  This will be useful once mouse wheel zoom is supported.