Handle coarser timestamps when checking if a scan is up to date
Since some filesystems have above-millisecond timestamp
granularity and update them asynchronously it can happen that
the scan results file has the timestamp up to a second the one
of the map it's produced from (such as when the scan happens to
run really fast and falls within the same time slot). In this
case assume it's up to date instead of unnecessarily rescanning.