fix: Instruct the LLM to be briefer in the vector terms
I noticed sometimes the LLM tries to put a lot in the `instructions`
part of the vector terms response; this part is only meant to direct
reasoning and we're ultimately throwing it out, so it doesn't make
sense to be very verbose. It also appears to hit against some limits,
at least on Gemini.