chore: Control concurrency of GitHub Actions jobs
This change will automatically cancel any still running or pending
jobs from previous commits on a new push to a branch.

As a bonus, it will also cancel pending releases if something new
is pushed to main. This means if we merge a couple branches in
quick succession it won't lead to multiple releases getting pushed.