fix: Chat input now gets focus automatically
This solution doesn't require custom code editor integrations to gain
focus. The chat input will automatically update due to the logic in
`mounted`, triggering a focus when `updated` is called. It also binds to
the window's `onfocus` event handler property to restore focus if the
user switches to another window and returns.