fix: Normalize Windows path locations
In practice, Windows paths are represented in many different forms. Path
separators are interchangable and paths are case insensitive.
Additionally, paths constructed via a file scheme URI may contain a
leading slash, similar to an absolute path in POSIX.

This was preventing pinned items from being picked up in Windows. A
pinned file path may be something like:
```
/c:/Users/Administrator/Project/app.rb:2-24
```

This is problematic for a few reasons:
1. The leading slash corrupts the path and it is not usable. Ideally
   clients would not send us paths formatted in this manner, but it's
   already happening. This behavior is observed both from JetBrains and
   Visual Studio Code.
2. The path separators are valid, but don't match other cases in our
   code, such as when we compare the absolute file path against known
   project directories. In this case, known project directories will use
   the default separator (\\), so checks to see if a file starts with
   one of these paths will always fail.
3. A case inconsistency in the drive letter will also cause the above
   scenario to fail. Windows doesn't actually care, but our string
   comparisons obviously do.