fix: Limit number of subdirectory entries when listing files
Unlimited number of subdirectory entries can easily blow up context
length (eg. when one of the subdirectories is node_modules). Instead
if the number of entries is high (currently hardcoded to a limit of 100)
just pass the number of entries; the gatherer can request a directory
listing explicitly when needed (the limit doesn't apply to toplevel
entries).