refactor (wip): Token stream parsing abstraction
Aims to facilitate the extension of new token metadata types, such as
inline AppMap recommendations or porting the <change> XML blocks.

This strategy currently attempts to parse and emit events as soon as
possible, but I suspect we'll hit an issue eventually because we're not
buffering at least a full line of text before parsing. This makes it
difficult (if not impossible?) to differentiate between the start of a
new line and the middle of an existing line.

E.g., if we receive the following two tokens:
["hello", "```"]

Once the first token is processed and emitted, the second token has no
idea where it's placement is in the overall content.