Replace API calls with telemetry
API calls for creating conversations and messages with the deprecation of
remote Navie just function as glorified special telemetry.

These changes refactor the code to use standard telemetry instead. This
not only simplifies the code removing special cases, but also enables
corporate clients to use a custom telemetry backend to ingest this data.