refactor: Use the environment configuration to seed the model registry
There are two ways to configure an LLM backend: environment variables
or dynamic configuration via RPC. While the latter is meant to override
the former, the fallback (environment) configuration wasn't properly
visible through the interfaces for the dynamic configuration.

This change provides a unified view by seeding the model registry
with the fallback configuration as the default initial config.